Lower version requirements and refactor class properties
This commit lowers the minimum requirement for PHP, Symfony Framework Bundle and Symfony Twig Bundle in the composer.json file. This is a backward compatibility improvement. The other part of this commit includes refactoring: class properties are now declared and assigned in the constructor instead of using property promotion. This change helps to increase the code clarity and readability in GlideExtensionRuntime and GlideController classes.